Creation of Store
-toDecimal, and fromDecimal objects

These maps, representing the bases 2 through 36 are used for easy transition between a value and its representational character. For example, in base 11 'A' represents the value 10. This is the same as 9 representing 9, but since we write out numerical system in base 10 we don't have a single character for it.
This allows the program to make this relational logic without having a million if elses in the actual method logic.