Structure and method logic for buttons
The user is able to input a string representing any numerical value notated between base 2 and 36. The logic parses this string for the conversion, so creating a number field is unnecessary and actually detrimental since users must be able to use letters.
On clicking convert, JavaScript takes the string the user inputs and converts it character by character. The mathematical logic behind this is standard base conversion as there is no other way I have thought of to do this more simply.